Incrementally re-render results after `refetch` or `skip` when using `@defer` with `useSuspenseQuery` (#11035)
When issuing a query with an `@defer` directive, calling refetch or enabling a query by disabling the skip option would only re-render when the entire result was loaded. This felt like it defeated the purpose of the `@defer` directive in these cases.

This adds the ability to incrementally re-render results returned by `@defer` queries to match the behavior of the initial fetch. This also makes a first attempt at refactoring some of `QueryReference` to remove flag-based state and replace it with a status enum.

NOTE: This attempts to add support for incrementally re-rendering `fetchMore`, but there is currently a bug in core that prevents this from happening. A test has been added that documents the existing behavior for completeness.