Dispose of the `queryRef` when changing to previously used variables to allow the `InMemoryCache` to determine cached behavior (#11045)
When changing variables to a previously used set of variables, the fetch policy is now honored. Previously when using a fetch policy of `network-only` or `no-cache`, switching to previously used variables would return a cached result and would require a `refetch` to get data from the network again.

Now query refs are disposed of when changing variables so that switching back to previously used variables will either return from the `InMemoryCache` or will execute a network request to load from the server.