Adds support for `skipToken` in options for `useSuspenseQuery` and `useBackgroundQuery` and soft deprecates `skip` option (#11112)
Adds support for a new `skipToken` sentinel that can be passed as options to `useSuspenseQuery` and `useBackgroundQuery`. This works in place of `skip` for both of these hooks. As such, the `skip` option is now labeled as `deprecated` to nudge people to use this feature, which is generally more type-safe than the `skip` option.

This also changes the way `queryRef` is returned when `skip` is true or `skipToken` is used with `useBackgroundQuery`. Previously, a `queryRef` would always be returned, regardless of whether query execution was skipped. This has been updated to conditionally return a queryRef only when query execution is enabled. `undefined` is returned when initializing the hook skipped.

---------

Co-authored-by: Lenz Weber-Tronic <lorenz.weber-tronic@apollographql.com>