Avoid excessive `cache.watch` / unsubscribe on every render in `useFragment` (#11464)
The subscribe callback passed to `useSyncExternalStore` was not wrapped in a `useCallback` and because of this, `useFragment` will unsubscribe from the previous `cache.watch` and resubscribe with a `cache.watch` on every render. While `cache.watch` should be fairly low overhead, there is no need to have to unsubscribe/resubscribe on every render. Instead the `cache.watch` handler now lasts until either the cache or options have changed, which will then unsubscribe and rewatch with the new options.