Make `extensions` available in `merge` functions (#13058)
Add a new `extensions` option to `cache.write`, `cache.writeQuery`, and
`client.writeQuery` that makes them available to cache `merge`
functions. `extensions` returned in GraphQL operations that write to the
cache are now provided to the cache write operation.

This is the precursor to the change needed to handle a more robust
`refetch` with the `@stream` directive as discussed in
https://github.com/apollographql/apollo-client/pull/12923#discussion_r2518383075.

---------

Co-authored-by: jerelmiller <565661+jerelmiller@users.noreply.github.com>
Co-authored-by: Lenz Weber-Tronic <lorenz.weber-tronic@apollographql.com>
Co-authored-by: phryneas <4282439+phryneas@users.noreply.github.com>