Truncate `@stream` arrays only on last chunk by default (#13069)
Updates the default `@stream` behavior to maintain existing cache data
as much as possible. Currently the `@stream` array is truncated when the
first chunk arrives which means you might lose data on screen suddenly
while the data is fetched. We want to maintain the truncation, but only
when the server completes the stream of data.

---------

Co-authored-by: Lenz Weber-Tronic <lorenz.weber-tronic@apollographql.com>